refactor(api): split r7.ts into focused route modules
r7.ts was 3,563 lines (3,499 after the warm-session simplification) holding
nine unrelated concerns. Pure structural move along the seams that were
already there: no behaviour change, no route path change, no response-shape
change. tests/spec/routes.generated.json regenerates byte-identical (578
routes).

  group-grants.ts          322   group grants: list, attach, re-role, detach
  warm-sessions.ts         287   warm create + the deprecated claim
  project-sessions.ts      566   session create, list, read, share, patch, delete
  session-transcripts.ts   168   chat transcript + voice-call transcript
  project-audit.ts         458   project log, sandbox ingestion, session timeline
  approvals.ts             457   approval inbox, needs-input, resolve
  resource-grants.ts       331   per-agent/skill scoping
  session-scope.ts         705   connector/secret scope + model switch
  session-config.ts        280   config freshness + reload (JSON and streamed)

Every module is under the 1,000-line ceiling; the largest is session-scope.ts
at 705.

`parseBoundedPositiveInt` moved to lib/serializers.ts beside the other
request-parsing helpers — three modules need it now.

r7.ts is deleted rather than left as a barrel: a file that only re-exports its
own replacements is a thin wrapper. projects/index.ts registers the nine
modules directly. The four tests that read r7.ts as SOURCE now read the module
that owns the code they assert on, and r7-audit-ingestion.test.ts is renamed
to project-audit-ingestion.test.ts to match.
